Create an R Project File in your Academic Project

create_rproj() finishes by way of command line what Rstudio will do naturally. It will add a simple .Rproj file to your directory.

create_rproj(path, ...)

Arguments

path
the directory name for the project (e.g. "my-project", "dissertation", or whatever)

...
optional stuff, but you can ignore this

Value

create_rproj() creates an .Rproj file in your directory. It should be run in the command line after running create_project{} in the command line. You will NOT need to do this if you elect to use Rstudio's interface for the creation of a project. Rstudio will do that for you automatically.

Examples

if (FALSE) {
create_project("example") # creates new directory
create_rproj("example") # creates new \code{.Rproj} file in that directory
}