Render Your R Markdown Appendix to Various Outputs

render_appendix() takes various arguments, most effectively built into the function, and renders your R Markdown manuscript to various outputs supported by stevetemplates.

render_appendix(
  file = "appendix.Rmd",
  output_dir = "doc",
  outputs = c("pdf", "pdf-anon"),
  parameters,
  latex_engine = "xelatex",
  dev = "cairo_pdf"
)

Arguments

file
the name of the R Markdown file containing that is your appendix. Defaults to "appendix.Rmd".

output_dir
the output directory to contain the formatted manuscript. Defaults to "doc". Optimally, this is a subdirectory of the directory containing the manuscript. A user who wants the formatted manuscript to be in the same directory as the R Markdown file should specify output_dir = NULL here.

outputs
the various formatted manuscript types the user wants, supplied as a character vector. Must be one or more of "pdf", "pdf-anon", "word", and/or "html". No other formats are supported right now. Defaults are "pdf" and "pdf-anon".

parameters
optional parameters, specified as a character, based on what's in your R Markdown file, passed as params in the render() function in R Markdown. If no parameters are specified here, the function defaults these parameters to anonymous=TRUE, doublespacing=TRUE, which is then wrapped in a list to be passed to the params argument in render(). Do note this primarily concerns the anonymous manuscript type. These are somewhat advanced-level arguments, so the user should be careful what they do here and should have a firm idea what they are doing here.

latex_engine
the LaTeX engine the user may want to use. Defaults to "xelatex". You can overwrite this if you would like, but why would you?

dev
the graphics device for LaTeX PDFs. Defaults to "cairo_pdf". You can overwrite this, but why would you?

Value

render_ms() takes various arguments, most effectively built into the function, and renders your R Markdown manuscript to various outputs supported by stevetemplates.

Examples

if (FALSE) {
render_appendix()
}